refactor(astro-config): fixed directory + config settings : Check comment in base.config.mjs
