--- config/astro/algorithms.config.mjs
@@ -1,10 +1,0 @@
-import { defineConfig } from "astro/config";
-import { baseConfig } from "./base.config";
-
-export default defineConfig({
-  ...baseConfig,
-  outDir: "./dist/algorithms",
-  build: {
-    format: "directory",
-  },
-});
