+++ site/config/postcss.config.ts
@@ -3,7 +3,7 @@ import type { Config } from "postcss-load-config";
 export default {
   plugins: {
     tailwindcss: {
-      config: "./config/tailwind/base.ts",
+      config: "./config/tailwind/tailwind.config.ts",
     },
     autoprefixer: {},
   },
